pyproject.toml
[tool.poetry]
name = "lifelogger"
version = "0.1.0"
description = ""
authors = ["John <tagmeh@gmail.com>"]
#readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2"
django-rest-framework = "^0.1.0"
djangorestframework = "^3.14.0"
django-filter = "^23.1"
markdown = "^3.4.3"
django-environ = "^0.10.0"
python-dotenv = "^1.0.0"
drf-spectacular = { extras = ["sidecar"], version = "^0.26.2" }
djangorestframework-simplejwt = "^5.2.2"
pytest-cov = "^4.0.0"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
mypy = "^1.2.0"
django-stubs = "^1.16.0"
pytest = "^7.3.1"
pytest-django = "^4.5.2"
djangorestframework-stubs = "^1.10.0"
pytest-factoryboy = "^2.5.1"
faker = "^18.4.0"
ipython = "^8.12.0"
coverage = "^7.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.mypy]
#PycharmProjects\LifeLogger\lifelogger> mypy --config-file=../pyproject.toml .
# Can't seem to run mypy from the same level as pyproject.toml, so we must specify the config in the command.
python_version = 3.11
disallow_untyped_defs = true
ignore_missing_imports = true
strict = true
exclude = ["tests"]
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]

[tool.django-stubs]
django_settings_module = "lifelogger.settings"

[tool.djangorestframework-stubs]
django_settings_module = "lifelogger.settings"


[tool.black]
line_length = 120
skip-string-normalization = true


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "lifelogger.settings"
python_files = "test_*.py"
addopts = "--cov --cov-report html --cov-report term-missing:skip-covered --cov-fail-under 95"

[too.coverage.report]
omit="tests"